requirements.txt
# main app 
Django==3.2.25
djangorestframework==3.14.0
drf-spectacular>=0.15.1,<0.16
Pillow>=9.0.0
djangorestframework-simplejwt>=4.7.2,<5.0.0
PyJWT>=2.0.0,<2.5.0
cryptography>=3.4.7
aiohttp==3.8.6
uvicorn==0.23.2
gunicorn==21.2.0
redis>=4.0.0
channels==3.0.5
celery==5.2.7
scipy==1.9.3
tqdm==4.65.0
psycopg2-binary>=2.9.3
django-celery-beat==2.2.1
django-environ

# elastic search 
django-elasticsearch-dsl==7.2.2
elasticsearch==7.13.4

# ai models 
faiss-cpu==1.7.4
numpy==1.23.5
torch>=1.12.0
sentence-transformers>=2.2.2
transformers>=4.35.0
scikit-learn==1.2.2
pandas==1.5.3